@@ -54,7 +54,7 @@ enum SignInStatus {
   success, // Estado cuando el inicio de sesión fue exitoso
   notFound,
   server,
-  failure, // Estado cuando ocurrió un error
+  failure, // , notFoundEstado cuando ocurrió un error
   invalidCredentials, // Estado cuando las credenciales son incorrectas
   networkError, // Estado cuando hay un problema de red
   unknown,
