@@ -285,11 +285,6 @@ class _LoginScreenState extends State<LoginScreen> {
                                             Text(AppText.invalidCredentials)),
                                   );
                                 } else if (state.signInStatus ==
-                                    SignInStatus.notFound) {
-                                  ScaffoldMessenger.of(context).showSnackBar(
-                                    const SnackBar(content: Text("Not Found")),
-                                  );
-                                } else if (state.signInStatus ==
                                     SignInStatus.networkError) {
                                   ScaffoldMessenger.of(context).showSnackBar(
                                     const SnackBar(
@@ -405,7 +400,7 @@ class _LoginScreenState extends State<LoginScreen> {
                             //     ),
                             //   ],
                             // ),
-                            
+
                             // no tienes cuenta
                             Row(
                               mainAxisAlignment: MainAxisAlignment.center,
