@@ -143,13 +143,13 @@ class CouponWidget extends StatelessWidget {
                                 'Descripcion Descuento en la compra de la segunda hamburguesa',
                                 style: textTheme.labelMedium!.copyWith(
                                   color: AppColors.primary,
-                                  fontWeight: FontWeight.w400,
-                                  fontFamily: 'onset',
+                                  //fontWeight: FontWeight.w400,
+                                  //fontFamily: 'onset',
                                   letterSpacing: 1,
                                   overflow: TextOverflow.ellipsis,
                                   height: 1.1,
                                 ),
-                                maxLines: 3,
+                                maxLines: 2,
                               ),
                             ],
                           ),
@@ -205,11 +205,11 @@ class CouponWidget extends StatelessWidget {
                       'Valido de 15/10/2024 hasta 28/10/2024',
                       style: textTheme.titleSmall!.copyWith(
                         color: AppColors.primary,
-                        fontWeight: FontWeight.w500,
+                        //fontWeight: FontWeight.w500,
                         letterSpacing: 1,
                         overflow: TextOverflow.ellipsis,
                         height: 1.1,
-                        fontSize: 10,
+                        fontSize: 8,
                       ),
                       maxLines: 2,
                     ),
