@@ -132,7 +132,7 @@ class CouponPage extends StatelessWidget {
                           ),
                         );
                       },
-                    )
+                    ),
                   ],
                 );
             }
