@@ -6,6 +6,9 @@ import 'package:puntos_smart_user/app/core/constants/app_text.dart';
 import 'package:puntos_smart_user/app/core/extension/custom_extension.dart';
 import 'package:puntos_smart_user/app/features/auth_feature/presentation/widgets/custom_button_widget.dart';
 
+import '../../../../core/theme/app_colors.dart';
+import '../../../../core/widgets/custom_arrow_back.dart';
+
 class FavoritesScreen extends StatefulWidget {
   const FavoritesScreen({super.key});
 
@@ -59,29 +62,43 @@ class _FavoritesScreenState extends State<FavoritesScreen> {
                   SliverAppBar(
                     pinned: true,
                     backgroundColor: Colors.white,
-                    leading: InkWell(
-                      onTap: () {
-                        context.pop();
-                      },
-                      child: const Icon(
-                        Icons.arrow_back_ios_new,
-                        color: Colors.black54,
+                    // leading: InkWell(
+                    //   onTap: () {
+                    //     context.pop();
+                    //   },
+                    //   child: const Icon(
+                    //     Icons.arrow_back_ios_new,
+                    //     color: Colors.black54,
+                    //   ),
+                    // ),
+                    leadingWidth: 70,
+                    leading: const Padding(
+                      padding: EdgeInsets.only(left: 16, bottom: 4),
+                      child: CustomButtonArrowBack(),
+                    ),
+                    centerTitle: true,
+                    title: Text(
+                      AppText.myFavorities,
+                      style: textTheme.bodyLarge!.copyWith(
+                        color: AppColors.blacknew,
+                        fontWeight: FontWeight.w600,
                       ),
+                      textAlign: TextAlign.center,
                     ),
                   ),
-                  SliverToBoxAdapter(
-                    child: Padding(
-                      padding: const EdgeInsets.symmetric(
-                          horizontal: 16, vertical: 12),
-                      child: Text(
-                        AppText.myFavorities,
-                        style: textTheme.titleLarge!.copyWith(
-                          color: Colors.black87,
-                          fontWeight: FontWeight.w600,
-                        ),
-                      ),
-                    ),
-                  ),
+                  // SliverToBoxAdapter(
+                  //   child: Padding(
+                  //     padding: const EdgeInsets.symmetric(
+                  //         horizontal: 16, vertical: 12),
+                  //     child: Text(
+                  //       AppText.myFavorities,
+                  //       style: textTheme.titleLarge!.copyWith(
+                  //         color: Colors.black87,
+                  //         fontWeight: FontWeight.w600,
+                  //       ),
+                  //     ),
+                  //   ),
+                  // ),
                   SliverToBoxAdapter(
                     child: SizedBox(
                       height: 50,
