@@ -2,6 +2,10 @@ import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 import 'package:puntos_smart_user/app/features/auth_feature/presentation/widgets/custom_button_widget.dart';
 
+import '../../../../core/constants/app_text.dart';
+import '../../../../core/theme/app_colors.dart';
+import '../../../../core/widgets/custom_arrow_back.dart';
+
 class PersonalPreferencesScreen extends StatefulWidget {
   const PersonalPreferencesScreen({super.key});
 
@@ -47,14 +51,28 @@ class _PersonalPreferencesScreenState extends State<PersonalPreferencesScreen> {
                   SliverAppBar(
                     pinned: true,
                     backgroundColor: Colors.white,
-                    leading: InkWell(
-                      onTap: () {
-                        context.pop();
-                      },
-                      child: const Icon(
-                        Icons.arrow_back_ios_new,
-                        color: Colors.black54,
+                    // leading: InkWell(
+                    //   onTap: () {
+                    //     context.pop();
+                    //   },
+                    //   child: const Icon(
+                    //     Icons.arrow_back_ios_new,
+                    //     color: Colors.black54,
+                    //   ),
+                    // ),
+                    leadingWidth: 70,
+                    leading: const Padding(
+                      padding: EdgeInsets.only(left: 16, bottom: 4),
+                      child: CustomButtonArrowBack(),
+                    ),
+                    centerTitle: true,
+                    title: Text(
+                      AppText.personalPreferences,
+                      style: textTheme.bodyLarge!.copyWith(
+                        color: AppColors.blacknew,
+                        fontWeight: FontWeight.w600,
                       ),
+                      textAlign: TextAlign.center,
                     ),
                   ),
                   SliverToBoxAdapter(
